Refactor more files
